refactor: move Flattener out of transform module
The Flattener struct is not used a single time in the transforms module,
so it doesn't make sense to define it there ... the module is already
long enough.
